twilio/requirements.txt
flask==2.3.3
twilio==8.8.0
requests==2.31.0
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
